Omap4 : Enabling SGX driver initialization
Entries added to the init script to

1. set the SGX driver permissions (pvrsrvkm)
2. run the pvrsrvinit binary which brings the driver to a fully
   initialized state.

This change is needed to enable SGX driver that will soon be an
integral part of the kernel. Note that if the SGX driver is not
part of the kernel (i.e. not compiled in) or pvrsrvinit binary
doesn't exist this change has no effect and everything works as
before.

Change-Id: Ifb0903aedd012fabb8e7930fca02b43ae6d63eab