Blaze sensor HAL: Add pressure and temperature sensor support
Adding the pressure and temperature support from the BMP085 pressure sensor

This sensor can report both the pressure (in Pascals) and temperature (in Celcius).
Since this sensor can support both the HAL needs to ensure that the sensor continues to report
any open sensors.

For instance is the pressure and temperature are both being reported and the pressure is no longer
needed then the HAL must keep the temperature reporting until the applciation requests it to be off.

Change-Id: Ied4666d6bc090bf81fb5c5d92b3aaa22a6db9b8a
Signed-off-by: Dan Murphy <dmurphy@ti.com>