Blaze sensor HAL: Add pressure and temperature sensor support
Adding the pressure and temperature support from the BMP085 pressure sensor

This sensor can report both the pressure (in Pascals) and temperature (in Celcius).
Since this sensor can support both the HAL needs to ensure that the sensor continues to report
any open sensors.

For instance is the pressure and temperature are both being reported and the pressure is no longer
needed then the HAL must keep the temperature reporting until the applciation requests it to be off.

Change-Id: I69ca6bd3d416af0f4af32e8f34f471eb655ec422
Signed-off-by: Dan Murphy <dmurphy@ti.com>

Patch Set 2: rebased on current trees.  No changes made.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>