Blaze: Enable RenderScript from the Product Definition makefile
This patch enables the following RenderScript modules (libRS.so and librs_jni.so) as part
of the Product Definition makefile. This patch deprecates this other patch: http://review.omapzoom.org/#change,4425
which does the same thing but incorrectly. This change also only affects the p-froyo branch since we are not
enabling RenderScript in p-eclair.

Change-Id: I7f35f390faa9e0cdd9b531aa5abbc3929072faf9
Signed-off-by: Carlos Santa <c-santa@ti.com>