omap4: board: change mmc registration order
eMMC is a fixed device always available on omap4 boards.
So make eMMC as /dev/mmcblk0
and
External mmc as /dev/mmcblk1

Now the bootarg for external mmc has to be:
root=/dev/mmcblk1p2

This solves many issues we have previously seen with detachable mmc number.
We needed to boot with external MMC card attached for eMMC booting to work.
With this fix, that limitation is gone.

Change-Id: Id60c983882e2de50a9743c2200987126ef55e15f
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>
Signed-off-by: Chikkature Rajashekar, Madhusudhan <madhu.cr@ti.com>