add audio properties for mic and pwr mode preference
omap.audio.mic.xyz=AMicXYZ will configure the system to use
only analog mics for either main or sub mic.
this is the same as the current
implementation and serves as the default value.

omap.audio.mic.xyz=DMicXYZ will configure the audio input to use
digital mics for the default audio input

omap.audio.power=HQ will be the High Quality mode.
This is the same as the legacy implementation with default
output sample rate of 48kHz.

omap.audio.power=LP will be used to configure the system
to run in the new Low Power mode using the output sample rate
of 44.1kHz and some other optimizations as required.

Change-Id: Ifea432def6a730329833f0f4b40ceb263293696e
Signed-off-by: Chris Kelly <c-kelly@ti.com>