device/ti/blaze: init.rc: change uim-rfkill path
Based on a customer enhancement request the uim-rfkill was moved
to /system/bin/ instead of the previous /system/xbin/.
NOTE:
This change is dependent on the uim build for release change which
modifies the uim's Android.mk for executable's location.

Change-Id: I346c1e888fc172f327934eadffca44de55c51497
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>

Recreated this patch since the database got corrupted for the
original patch:  8158.

Signed-off-by: Jean Johnson <jean-johnson@ti.com>