Change permission on scaling governor to apply boost api updates at run time
During browsing scenario, in order to achieve the best performance it is necessary to disable the hotplug governor and enable the
performance governor. Also in multi-core architecture, we need to enable all the cores. This change ensures that appropriate file
permissions are given in order to make these run time updates. Note the browser application will revert back to hotplug governor,
once the browser application crashes or pushed to background or low memory scenario or application is closed.

Change-Id: I3c023399a2c8cfc801474d979bc7b30b1e4fa3cd