GingerBread Sensor HAL: Initial release of Sensor HAL
This is the initial release of the Sensor HAL for Ginger Bread.
Added the base class, input reader and other sensor files
Currently only the Accelerometer is supported with this offering.

Add the proximity and light sensor for now so that the changes up coming
will be minimal but the code will need to be modified for the blaze HW.

Removed old sensor HAL implementation

Change-Id: Ibfdc48194ff0a45ccbf72d9a6615a0587f3148b9
Signed-off-by: Dan Murphy <dmurphy@ti.com>