Name of sys M3 binary changed to base_image_sys_m3.xem3
Description: Name of sys M3 binary changed to base_image_sysm3.xem3.
This is to align with ducati naming convention for the binaries.
This patch has to be taken when we move to 1.21-p5 ducati, because the
corresponding change on ducati is present in 1.21-p5.

Validation# tested AV playback with 1.21-p5 ducati binaries.

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch set 3: Cherry-pick on current tree.  No changes made.
Taking verifications, approvals and build test from patch set
2.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: I19c79569a58696232c9f367b2422653a8140f0cc