Initial blaze audio hal hw module
This is the initial contribution of the primary blaze audio hw module.
This supports playback on headset and handfree with routing for only
AUDIO_DEVICE_OUT_WIRED_HEADSET and AUDIO_DEVICE_OUT_SPEAKER.

This was orignally copied from a draft of the initial
samsung/tuna version and refactored to blaze. It is the intent to keep
this blaze module as a superset of both the panda and tuna modules which
each use a subset of the audio features available on blaze.

Signed-off-by: a0271430 <a0271430@ti.com>
Signed-off-by: Chris Kelly <c-kelly@ti.com>

Patch Set 9:  Simple cherry-pick on current tree.  No changes
made.  Taking verification, approvals and build test from patch
set 8.
Signed-off-by: Jean Johnson <a0271255@ti.com>
Change-Id: I14a2d576254944447376ea8cacc80209fb5232b4