[OMAP4] Audio: Workaround ABE microphone disable problem
The analog microphone uplink stops working if any
downlink route is changed when the modem PCMs are
open, so as a workaround, only modify the earpiece
route when the modem PCMs are closed.

Signed-off-by: Ricky Bhatia <a0866433@ti.com>

Patch Set 2: Cherry-pick on current tree. No changes made.
Taking verifications and approvals from previous patch set.
Signed-off-by: Jean Johnson <a0271255@ti.com>
Change-Id: I021750cc5737025af122f5ed4bb29d8e525d232e