[OMAP4] Fix audio output to play audio other than 44k stereo
When opening an output stream, the driver needs to return its
format, sample rate and channel mask. Failure to do so will
cause AudioFlinger to create the ouput as DIRECT, regardless of
the output flags.
The consequence of the output thread being in DIRECT mode is
that no mixing is done (one sound at a time), and only audio
that plays at the HAL's expected audio characteristics (44k
16bit stereo) can be played.
The fix consists of returning the format, chanel mask and sample
rate values when opening the output stream.

Signed-off-by: Ricky Bhatia <a0866433@ti.com>

Patch Set 2: Cherry-pick on current tree. No changes made.
Taking verifications and approvals from previous patch set
Signed-off-by: Jean Johnson <a0271255@ti.com>

Change-Id: Ib1efbac47408bd97502ef7616af5a8040a13565c