audio: fix latency and buffer size calculation
out_get_buffer_size() was returning the total buffer size in frames
instead of the period size in bytes. It should also take the
resampling into account so that the audio flinger buffer duration
somehow matches the period duration.

The calculation in out_get_latency() was assuming the period size
in the pcm config structure is in bytes whereas it is in frames.

Signed-off-by: Priyesh Bisla <a0271372@ti.com>

Patch set 2:  Simple cherry-pick on current tree.  No changes
made.  Taking approvals and verifications from patch set 1.
Signed-off-by: Jean Johnson <a0271255@ti.com>
Change-Id: I95aaa0aece9a324d12c455c3e81d8d64326432de