audio: reselect output device when changing mode
It's necessary to reselect the output device when changing from
the AUDIO_MODE_IN_CALL state to AUDIO_MODE_NORMAL, otherwise the
ASoC driver cannot find a route from the PCM to the endpoint.
This is possibly a workaround for a bug in the ASoC driver.

Also, this change means the audio HAL will no longer disable the
multimedia mixer routes when in a call. This is because audio
can be played during a call.

Signed-off-by: Priyesh Bisla <a0271372@ti.com>

Patch Set 2:  Simple rebase on current tree.  No changes made.
Taking approvals and verification from patch set 1.
Signed-off-by: Jean Johnson <a0271255@ti.com>
Change-Id: Iade3f110bcd6c84e97587c9c58bab1d71ce8fdc9