over-ride Bluetooth default feature
Hot-On/Off is a new feature brought in, which enables hci0 to be UP and
RUNNING all the time, and Bluetooth Off would only mean the PSCAN and ISCAN
mode is removed for the adapter and the connections are removed.

We don't really want this, since our Bluetooth Adapter doesn't take really
long to turn on or load the firmware. Moreover we would want our UART ports to
be closed when Bluetooth is turned off.

Signed-off-by: Pavan Savoy <x0099669@ti.com>