Blaze: Adding proper permission for DCC files
[Comment]
Adds correct permissions for DCC binary folders so that
the cameraHAL can access it without any issues.

Some info about the DCC files relocation:

The DCC lookup path is changed in the domx, to a path
which is not mounted as a read-only as default.
So DCC binaries are moved from /system/etc/omapcam/
to /data/misc/camera/.

This would enable to dynamicaly change the dcc settings,
without needing to remount the /system/ partition.

[DR]
-

Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>
Change-Id: I1d92a8c4f8b3f32bcebc578a445b9de02c14f4f4