fastboot.sh: Fix resizing userdata.img
If the label option given to make_ext4fs (-a)
does not match the directory name, then the
make_ext4fs util is changing the permissions of
all files and folders to root when creating new
image. So changing name of temporary directory
we create and label option.

Adding '-o grpid' option for mount to keep group id
of the directory instead of using fsgid of the current
process.

Also, resizing userdata.img requires sudo privileges
so, if anyone executes the command as a regular user
we were corrupting the userdata.img. So, in this case,
just warn users and flashing the original userdata.img.

Change-Id: I8e5559faae7209abd4789fbf03d4f24772db0366
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>