OMAP5sevm: Add bluetooth
Bluetooth to work on OMAP5 sevm  over the shared transport requires the UIM
to be started via the init.rc.
Also setting right the permission for the UIM service which is started as
user bluetooth.
UIM needs access to the UART connected to BT/FM/GPS/NFC and hence also
modifying the owner/group of the UART.
This patch also sets right the permission of the rfkill entry created by
bluetooth kernel sub-system which is accessed by bluedroid which belongs
to group bluetooth.

Change-Id: I03220179bde11a1c72003dcdb196d4e21e8ba80c
Signed-off-by: Shitiz Kumar <shitiz_kumar@ti.com>