GPS: Create an entry for starting devproxy.
Create an entry for starting devproxy service at bootup.
devproxy service is required by the GPS host software based
on TI BT/FM/GPS combo chipset.

The device-proxy is a background agent or daemon that acts on behalf
of the GNSS receiver to provision a device interface to a client application.
This component, dev-proxy, presents the services and features of the GNSS
device in the form of discrete API(s) to the application world.

Change-Id: I7e8e7a46d3e232cef222e9f396a35d2738ac1818
Signed-off-by: Satish Chandra <x0141690@ti.com>