keypad: define omap4-keypad as special-function keyboard
As per Android definition: a special-function keyboard
is only used to perform system control functions rather
than for typing. The omap4-keypad device is omap5 sEVM
platform is used only for the following system control
functions:

SEARCH | HOME | BACK | VOLUME

Note that special-function keyboard will never be registered
as built-in keyboard. This is because a special-function
keyboard is by definition not intended to be used for general
purpose typing.

Change-Id: Ib833525dcd1c562138de05c69b7be382a8ceb006
Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>