OMAP5:sEVM: Add symlink for K3.4 mount compatibility
Prior to Linux Kernel 3.4, the omap_hsmmc platform
device sysfs entry was at
/sys/devices/platform/omap/omap_hsmmc.X
and so, the init process created a softlink to mount
device at
/dev/block/platform/omap/omap_hsmmc.X

But in K3.4, the omap_hsmmc platform device sysfs
entry is now at
/sys/devices/platform/omap_hsmmc.X
and so, the init process created softlink to mount
device at
/dev/block/platform/omap_hsmmc.X

So, we create a symbolic link so that mount succeeds.
Now, the same init.omapXXXX.rc will work both for
K3.4 and also for older kernels.

For older kernels, this symlink command will just fail
and mounts will succeed.

Change-Id: Icaa64de8dca1c390a866f70586a78ee928329f6e
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>