OMAP5: Removes the black area in FaceUnlock camera preview.
A black area is appeared at the left side of the camera
preview during the Face unlock with PATTERN unlock.
This issue is not noticed with Face unlock with PIN unlock.

This patch removes that black area and shows the correct
preview window.
This is fixed by changing the faceLockAreaView equal to the
one which is present in keyguard_screen_password_portrait.xml.

Change-Id: I56622114fa9efbf43040052bf9fbe400b0c30d41
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>