OMAP5sevm: Modem: Restart rild and media upon property update
'media' service is restarted upon an update of modem.audio property.
This is necessary to ensure proper audio RIL configuration.

'ril-daemon' service is restarted upon an update of rild.libpath.
This is necessary for the ril daemon to load the right vendor RIL.

modem-detect.sh is executed at early-boot to get the modem detected
earlier in the boot process and avoid useless restarts.

Change-Id: I9d1f234de8d159ae8973807cbc908d92a45da814
Signed-off-by: Patrick Combes <p-combes@ti.com>