init.omap5sevm.rc: Convert services to class main
Encryption was failing because some services are not allowing /data
to be unmounted.

core services are never shut down after starting
main services are shut down and restarted after the framework is restarted
late_start services are not started until after /data has been decrypted
and mounted

Thus if a service is a core service and is using a file on /data,
unmounting data will fail while trying encryption.

Change-Id: I23044249d6328e47b571ea3807bc92bf69f94abf
Signed-off-by: Dan Murphy <dmurphy@ti.com>