OMAP5: Enable auto-suspend for Ethernet on EHCI bus
The Ethernet controller connected to EHCI bus will prevent the CPU idle
as it always keeps the EHCI bus active resulting in USB Host 'ON' all the time,
even though the Ethernet cable is not connected. So enabling auto-suspend for
the Ethernet will make sure when there is no activity on the Ethernet the EHCI
controller can put the Ethernet to suspend state.

Change-Id: I0eae93175f462ddb69edc6f41ddfa4ca8f1963a5
Signed-off-by: Mantesh Sarasetti <mantesh@ti.com>
Cc: Moiz Sonasath <m-sonasath@ti.com>