OMAP5sevm: Reduce Sensor HAL init time
Reduce the sensor HAL init time by 1 second.

This patch creates an array of dev names and dev nodes that
is used to find the data node for interesting devices.

The reduction comes in the fact that the sensor HAL does not need
to iterate through the event nodes to find the interested device.
The array which is populated at init contains all devices that the sensor
HAL may be interested in.

When creating the sensor object the data file descriptor pointer will be passed
to the object.  This file descriptor will be used to read the data from the node.

Change-Id: I74cbed52ae1dfea3cf7e12035da7690e3945bb81
Signed-off-by: Dan Murphy <dmurphy@ti.com>