omap5sevm: enable screen wakeup on charger/usb plug/unplug
In previous android releases, the screen was waking up on
charger/usb plug/unplug events.
In latest jellybean sources google became more flexible and
added ability to change this behavior. By default this feature
was disabled and we need to enable it manually,

This patch enables it using android overlay technique like
it has been done for other GEDs.

Change-Id: Ic0be6917f9faf14757963210ae9f7ffa8c064302
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>